Header moved back to App for now, and working. Play handling moved to App from thumbnail for now.
